Break Recommendations out to separate file
Also eliminates an unnecessary dependency on Projection data
